% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils_not_exported.R
\name{convert_to_sf}
\alias{convert_to_sf}
\title{Converts from data frame to sf}
\usage{
convert_to_sf(data, crs, crs_column, x_geom)
}
\arguments{
\item{data}{a tibble or data frame}

\item{crs}{The coordinates reference system of the data. Specify if the data
doesn't have a \code{crs_column}, and you know the CRS.}

\item{crs_column}{a character string of length one specifying the column
storing the CRS (created automatically by \code{\link{ddbs_write_vector}}).
Set to \code{NULL} if absent.}

\item{x_geom}{name of geometry column}
}
\value{
sf
}
\description{
Converts a table that has been read from DuckDB into an sf object.
Optimized to handle binary WKB streams using wk and geoarrow.
}
\keyword{internal}
