% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_not_exported.R
\name{convert_to_sf_wkb}
\alias{convert_to_sf_wkb}
\title{Converts from data frame to sf using WKB conversion}
\usage{
convert_to_sf_wkb(data, crs, x_geom)
}
\arguments{
\item{data}{a tibble or data frame}

\item{crs}{The coordinates reference system}

\item{x_geom}{name of geometry column}
}
\value{
sf
}
\description{
Converts a table that has been read from DuckDB into an sf object.
}
\keyword{internal}
