% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_not_exported.R
\name{crs_equal}
\alias{crs_equal}
\title{Compare two CRS objects for equality}
\usage{
crs_equal(crs1, crs2)
}
\arguments{
\item{crs1}{First CRS object}

\item{crs2}{Second CRS object}
}
\value{
Logical indicating if CRS are equal
}
\description{
Properly compares CRS objects, handling different representations of the same CRS.
}
\keyword{internal}
