% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_extension.R
\name{ddbs_default_conn}
\alias{ddbs_default_conn}
\title{Get or create default DuckDB connection with spatial extension installed and loaded.}
\usage{
ddbs_default_conn(create = TRUE, ...)
}
\arguments{
\item{create}{Logical. If TRUE and no connection exists, create one.
Default is TRUE.}

\item{...}{Additional parameters to pass to \code{ddbs_create_conn()}}
}
\value{
A \code{duckdb_connection} or NULL if no connection exists and
create = FALSE
}
\description{
Get or create default DuckDB connection with spatial extension installed and loaded.
}
\details{
The connection is created internally when the first ckspatial  function is run. Every
function of the package runs on this connection when the argument \code{conn} is \code{NULL}.
}
\examples{
\dontrun{
## load package
library(duckspatial)

# get/create the default connection
conn <- ddbs_default_conn()

}
}
