% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbs_ops_unary.R
\name{ddbs_drop_geometry}
\alias{ddbs_drop_geometry}
\title{Drop geometry column from a duckspatial_df object}
\usage{
ddbs_drop_geometry(x)
}
\arguments{
\item{x}{Input spatial data. Can be:
\itemize{
\item A \code{duckspatial_df} object (lazy spatial data frame via dbplyr)
\item An \code{sf} object
\item A \code{tbl_lazy} from dbplyr
\item A character string naming a table/view in \code{conn}
}
Data is returned from this object.}
}
\value{
A lazy tibble backed by dbplyr
}
\description{
Removes the geometry column from a \code{duckspatial_df} object, returning a
lazy tibble without spatial information.
}
\examples{
\dontrun{
## load package
library(duckspatial)

## read data
countries_ddbs <- ddbs_open_dataset(
  system.file("spatial/countries.geojson",
  package = "duckspatial")
)

## drop geometry column
countries_tbl <- ddbs_drop_geometry(countries_ddbs)
}
}
