% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckspatial_df_sf_methods.R
\name{ddbs_format_crs}
\alias{ddbs_format_crs}
\title{Format a CRS object compactly for printing}
\usage{
ddbs_format_crs(crs)
}
\arguments{
\item{crs}{An sf crs object}
}
\value{
A character string
}
\description{
Format a CRS object compactly for printing
}
\keyword{internal}
