% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckspatial_df_sf_methods.R
\name{ddbs_geom_col}
\alias{ddbs_geom_col}
\title{Get the geometry column name}
\usage{
ddbs_geom_col(x)
}
\arguments{
\item{x}{A duckspatial_df object}
}
\value{
Character string with geometry column name
}
\description{
Get the geometry column name
}
