% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbs_ops_unary.R
\name{ddbs_is_simple}
\alias{ddbs_is_simple}
\title{Check if geometries are simple}
\usage{
ddbs_is_simple(
  x,
  conn = NULL,
  name = NULL,
  new_column = NULL,
  crs = NULL,
  crs_column = "crs_duckspatial",
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{An \code{sf} spatial object. Alternatively, it can be a string with the
name of a table with geometry column within the DuckDB database \code{conn}.
Data is returned from this object.}

\item{conn}{A connection object to a DuckDB database. If \code{NULL}, the function
runs on a temporary DuckDB database.}

\item{name}{A character string of length one specifying the name of the table,
or a character string of length two specifying the schema and table
names. If \code{NULL} (the default), the function returns the result as an
\code{sf} object}

\item{new_column}{Name of the new column to create on the input data. If NULL,
the function will return a vector with the result}

\item{crs}{The coordinates reference system of the data. Specify if the data
doesn't have a \code{crs_column}, and you know the CRS.}

\item{crs_column}{a character string of length one specifying the column
storing the CRS (created automatically by \code{\link{ddbs_write_vector}}).
Set to \code{NULL} if absent.}

\item{overwrite}{Boolean. whether to overwrite the existing table if it exists. Defaults
to \code{FALSE}. This argument is ignored when \code{name} is \code{NULL}.}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
a vector, an \code{sf} object with simplicity information or \code{TRUE} (invisibly) for table creation
}
\description{
Checks if geometries are simple (no self-intersections) from a DuckDB table using the spatial extension.
Returns the result as an \code{sf} object with a boolean simplicity column or creates
a new table in the database.
}
\examples{
\dontrun{
## load packages
library(duckspatial)
library(sf)

# create a duckdb database in memory (with spatial extension)
conn <- ddbs_create_conn(dbdir = "memory")

## read data
argentina_sf <- st_read(system.file("spatial/argentina.geojson", package = "duckspatial"))

## store in duckdb
ddbs_write_vector(conn, argentina_sf, "argentina")

## check simplicity
ddbs_is_simple("argentina", conn)

## check simplicity without using a connection
ddbs_is_simple(argentina_sf)
}
}
