% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbs_options.R
\name{ddbs_sitrep}
\alias{ddbs_sitrep}
\title{Report duckspatial configuration status}
\usage{
ddbs_sitrep()
}
\value{
Invisibly returns a list with the current status configuration.
}
\description{
Displays useful information about the current configuration, including global options
and the status of the default DuckDB connection.
}
\examples{
ddbs_sitrep()
}
