% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_not_exported.R
\name{ddbs_temp_table_name}
\alias{ddbs_temp_table_name}
\title{Generate unique temporary table name}
\usage{
ddbs_temp_table_name()
}
\value{
Character string with unique table name
}
\description{
Creates a unique name for temporary tables to avoid collisions.
}
\keyword{internal}
