% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_not_exported.R
\name{ddbs_temp_view_name}
\alias{ddbs_temp_view_name}
\title{Generate unique temporary view name}
\usage{
ddbs_temp_view_name()
}
\value{
Character string with unique view name
}
\description{
Creates a unique name for temporary views to avoid collisions.
}
\keyword{internal}
