% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckspatial-package.R
\docType{package}
\name{duckspatial-package}
\alias{duckspatial}
\alias{duckspatial-package}
\title{duckspatial: R Interface to 'DuckDB' Database with Spatial Extension}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fast & memory-efficient functions to analyze and manipulate large spatial data data sets. It leverages the fast analytical capabilities of 'DuckDB' and its spatial extension (see \url{https://duckdb.org/docs/stable/core_extensions/spatial/overview}) while maintaining compatibility with R’s spatial data ecosystem to work with spatial vector data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cidree.github.io/duckspatial/}
  \item \url{https://github.com/Cidree/duckspatial}
  \item Report bugs at \url{https://github.com/Cidree/duckspatial/issues}
}

}
\author{
\strong{Maintainer}: Adrián Cidre González \email{adrian.cidre@gmail.com} (\href{https://orcid.org/0000-0002-3310-3052}{ORCID})

Authors:
\itemize{
  \item Adrián Cidre González \email{adrian.cidre@gmail.com} (\href{https://orcid.org/0000-0002-3310-3052}{ORCID})
  \item Egor Kotov \email{kotov.egor@gmail.com} (\href{https://orcid.org/0000-0001-6690-5345}{ORCID})
  \item Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})
}

}
\keyword{internal}
