% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_not_exported.R
\name{feedback_query}
\alias{feedback_query}
\title{Feedback for query success}
\usage{
feedback_query(quiet)
}
\arguments{
\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
cli message
}
\description{
Feedback for query success
}
\keyword{internal}
