% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils_not_exported.R
\name{get_geom_name}
\alias{get_geom_name}
\title{Get column names in a DuckDB database}
\usage{
get_geom_name(conn, x, rest = FALSE)
}
\arguments{
\item{conn}{A DuckDB connection}

\item{x}{name of the table}

\item{rest}{whether to return geometry column name, of the rest of the columns}
}
\value{
name of the geometry column of a table
}
\description{
Get column names in a DuckDB database
}
\keyword{internal#'}
