% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_not_exported.R
\name{get_nrow}
\alias{get_nrow}
\title{Get the number of rows in a table}
\usage{
get_nrow(conn, table)
}
\arguments{
\item{conn}{A \code{DBIConnection} object to a DuckDB database}

\item{table}{name of the table}
}
\value{
number of rows in the table
}
\description{
Get the number of rows in a table
}
\keyword{internal}
