% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils_not_exported.R
\name{get_query_list}
\alias{get_query_list}
\title{Get names for the query}
\usage{
get_query_list(x, conn)
}
\arguments{
\item{x}{sf or character}

\item{conn}{A connection object to a DuckDB database. If \code{NULL}, the function
runs on a temporary DuckDB database.}
}
\value{
list with fixed names
}
\description{
Get names for the query
}
\keyword{internal}
