% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_not_exported.R
\name{get_query_name}
\alias{get_query_name}
\title{Get names for the query}
\usage{
get_query_name(name)
}
\arguments{
\item{name}{table name}
}
\value{
list with fixed names
}
\description{
Get names for the query
}
\keyword{internal}
