% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_not_exported.R
\name{get_st_predicate}
\alias{get_st_predicate}
\title{Gets predicate name}
\usage{
get_st_predicate(predicate)
}
\arguments{
\item{predicate}{A geometry predicate function. Defaults to \code{intersects},
a wrapper of \code{ST_Intersects}. See details for other options.}
}
\value{
character
}
\description{
Gets a full predicate name from the shorter version
}
\keyword{internal}
