% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckspatial_df.R
\name{is_duckspatial_df}
\alias{is_duckspatial_df}
\title{Check if object is a duckspatial_df}
\usage{
is_duckspatial_df(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if object is a duckspatial_df
}
