% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckspatial_df.R
\name{new_duckspatial_df}
\alias{new_duckspatial_df}
\title{Create a duckspatial lazy spatial data frame}
\usage{
new_duckspatial_df(
  x,
  crs = NULL,
  geom_col = NULL,
  source_table = NULL,
  source_conn = NULL,
  create_view = FALSE
)
}
\arguments{
\item{x}{Input: tbl_duckdb_connection, tbl_lazy, or similar dbplyr object}

\item{crs}{CRS object or string}

\item{geom_col}{Name of geometry column (default: "geom")}

\item{source_table}{Name of the source table if applicable}

\item{source_conn}{Name of the source connection if applicable}

\item{create_view}{Logical. If TRUE, creates a temporary view for the input query.
Otherwise it generates a temporary table.}
}
\value{
A duckspatial_df object
}
\description{
Extends tbl_duckdb_connection with spatial metadata (CRS, geometry column).
}
\keyword{internal}
