% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_not_exported.R
\name{overwrite_table}
\alias{overwrite_table}
\title{Feedback for overwrite argument}
\usage{
overwrite_table(x, conn, quiet, overwrite)
}
\arguments{
\item{x}{table name}

\item{conn}{A \code{DBIConnection} object to a DuckDB database}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}

\item{overwrite}{Boolean. whether to overwrite the existing table if it exists. Defaults
to \code{FALSE}. This argument is ignored when \code{name} is \code{NULL}.}
}
\value{
cli message
}
\description{
Feedback for overwrite argument
}
\keyword{internal}
