% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoarrow_methods.R
\name{as_nanoarrow_array_stream.duckspatial_df}
\alias{as_nanoarrow_array_stream.duckspatial_df}
\title{Convert a duckspatial_df to a nanoarrow_array_stream}
\usage{
as_nanoarrow_array_stream.duckspatial_df(x, ..., schema = NULL, native = FALSE)
}
\arguments{
\item{x}{A \code{duckspatial_df} object}

\item{...}{Additional arguments passed to downstream methods}

\item{schema}{Optional target schema for the entire stream.}

\item{native}{If TRUE, transforms WKB to a "Native" GeoArrow layout (e.g.,
Point, Polygon) using optimized Arrow-to-Arrow kernels. This layout is
optimized for high-performance rendering in tools like Deck.GL.}
}
\value{
A \code{nanoarrow_array_stream}
}
\description{
Convert a duckspatial_df to a nanoarrow_array_stream
}
