% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils_not_exported.R
\name{dbConnCheck}
\alias{dbConnCheck}
\title{Check if a supported DuckDB connection}
\usage{
dbConnCheck(conn)
}
\arguments{
\item{conn}{A DuckDB connection}
}
\value{
TRUE (invisibly) for successful import
}
\description{
Check if a supported DuckDB connection
}
\keyword{internal}
