% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbs_utils.R
\name{ddbs_create_conn}
\alias{ddbs_create_conn}
\title{Create a duckdb connection}
\usage{
ddbs_create_conn(dbdir = "tempdir")
}
\arguments{
\item{dbdir}{String. Either \code{"tempdir"} or \code{"memory"}. Defaults to \code{"tempdir"}.}
}
\value{
A \code{duckdb_connection}
}
\description{
Create a duckdb connection
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# load packages
library(duckspatial)

# create a duckdb database in disk
conn <- ddbs_create_conn(dbdir = "tempdir")

# create a duckdb database in memory
conn <- ddbs_create_conn(dbdir = "memory")
\dontshow{\}) # examplesIf}
}
