% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{ddbs_crs}
\alias{ddbs_crs}
\title{Check CRS of a table}
\usage{
ddbs_crs(conn, name, crs_column = "crs_duckspatial")
}
\arguments{
\item{conn}{a connection object to a DuckDB database}

\item{name}{a character string of length one specifying the name of the table,
or a character string of length two specifying the schema and table names.}

\item{crs_column}{a character string of length one specifying the column
storing the CRS (created automatically by \code{\link{ddbs_write_vector}})}
}
\value{
CRS object
}
\description{
Check CRS of a table
}
\examples{
1 + 1 # TODO

}
