% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_not_exported.R
\name{ddbs_default_conn}
\alias{ddbs_default_conn}
\title{Get or create default DuckDB connection with spatial extension installed and loaded}
\usage{
ddbs_default_conn(create = TRUE, ...)
}
\arguments{
\item{create}{Logical. If TRUE and no connection exists, create one.
Default is TRUE.}

\item{...}{Additional parameters to pass to \code{ddbs_create_conn()}}
}
\value{
A \code{duckdb_connection} or NULL if no connection exists and
create = FALSE
}
\description{
Get or create default DuckDB connection with spatial extension installed and loaded
}
\keyword{internal}
