% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{ddbs_drivers}
\alias{ddbs_drivers}
\title{Get list of GDAL drivers and file formats}
\usage{
ddbs_drivers(conn = NULL)
}
\arguments{
\item{conn}{A \code{DBIConnection} object to a DuckDB database.
If not specified (\code{conn = NULL}), uses the default connection created with \code{ddbs_default_conn()} (or a temporary one).}
}
\value{
\code{data.frame}
}
\description{
Get list of GDAL drivers and file formats
}
\examples{
\dontrun{
## load package
library(duckspatial)

## database setup
conn <- ddbs_create_conn()

## check drivers
ddbs_drivers(conn)
}
}
