% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_filter.R
\name{ddbs_filter}
\alias{ddbs_filter}
\title{Spatial Filter}
\usage{
ddbs_filter(conn, x, y, predicate = "intersection", crs = NULL)
}
\arguments{
\item{conn}{a connection object to a DuckDB database}

\item{x}{a table with geometry column within the DuckDB database. Data is returned
from this object}

\item{y}{a table with geometry column within the DuckDB database}

\item{predicate}{geometry predicate to use for filtering the data}

\item{crs}{the coordinates reference system of the data}
}
\value{
an sf object
}
\description{
Filters data spatially based on a spatial predicate
}
\examples{
1 + 1 ## TODO

}
