% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbs_ops_unary.R
\name{ddbs_geometry_type}
\alias{ddbs_geometry_type}
\title{Returns the geometry type of input features}
\usage{
ddbs_geometry_type(x, conn = NULL, by_feature = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{Input spatial data. Can be:
\itemize{
\item A \code{duckspatial_df} object (lazy spatial data frame via dbplyr)
\item An \code{sf} object
\item A \code{tbl_lazy} from dbplyr
\item A character string naming a table/view in \code{conn}
}
Data is returned from this object.}

\item{conn}{A connection object to a DuckDB database. If \code{NULL}, the function
runs on a temporary DuckDB database.}

\item{by_feature}{Boolean. If \code{TRUE} (default), returns the geometry type for
each feature. If \code{FALSE}, returns a single geometry type summary for
the entire dataset.}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
A factor with geometry type(s)
}
\description{
Returns the geometry type(s) from a \code{sf} object or a DuckDB table.
}
\examples{
\dontrun{
## load package
library(duckspatial)

## read data
countries_ddbs <- ddbs_open_dataset(
  system.file("spatial/countries.geojson", 
  package = "duckspatial")
)

# option 1: passing sf objects
# Get geometry type for each feature
ddbs_geometry_type(countries_ddbs)

# Get overall geometry type
ddbs_geometry_type(countries_ddbs, by_feature = FALSE)
}
}
