% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{ddbs_glimpse}
\alias{ddbs_glimpse}
\title{Check first rows of the data}
\usage{
ddbs_glimpse(conn, name, crs = NULL, crs_column = "crs_duckspatial")
}
\arguments{
\item{conn}{a connection object to a DuckDB database}

\item{name}{a character string of length one specifying the name of the table,
or a character string of length two specifying the schema and table names.}

\item{crs}{the coordinates reference system of the data. Specify if the data
doesn't have crs_column, and you know the crs}

\item{crs_column}{a character string of length one specifying the column
storing the CRS (created automatically by \code{\link{ddbs_write_vector}})}
}
\value{
\code{sf} object
}
\description{
Check first rows of the data
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## TODO
\dontshow{\}) # examplesIf}
}
