% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{ddbs_glimpse}
\alias{ddbs_glimpse}
\title{Check first rows of the data}
\usage{
ddbs_glimpse(conn, name, quiet = FALSE)
}
\arguments{
\item{conn}{A \code{DBIConnection} object to a DuckDB database}

\item{name}{A character string of length one specifying the name of the table,
or a character string of length two specifying the schema and table
names. If \code{NULL} (the default), the function returns the result as an
\code{sf} object}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
Invisibly \code{duckspatial_df} object
}
\description{
Prints a transposed table of the first rows of a DuckDB table, similarly
as the S3 \link[dplyr:glimpse]{dplyr::glimpse} method.
}
\examples{
\dontrun{
library(duckspatial)

# create a duckdb database in memory (with spatial extension)
conn <- ddbs_create_conn(dbdir = "memory")

## read data
argentina_sf <- ddbs_open_dataset(system.file("spatial/argentina.geojson", package = "duckspatial"))

## store in duckdb
ddbs_write_table(conn, argentina_sf, "argentina")

## glimpse the inserted table
ddbs_glimpse(conn, "argentina")
}
}
