% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbs_dimensions.R
\name{ddbs_has_dim}
\alias{ddbs_has_dim}
\alias{ddbs_has_z}
\alias{ddbs_has_m}
\title{Check geometry dimensions}
\usage{
ddbs_has_z(x, by_feature = FALSE, conn = NULL, quiet = FALSE)

ddbs_has_m(x, by_feature = FALSE, conn = NULL, quiet = FALSE)
}
\arguments{
\item{x}{Input spatial data. Can be:
\itemize{
\item A \code{duckspatial_df} object (lazy spatial data frame via dbplyr)
\item An \code{sf} object
\item A \code{tbl_lazy} from dbplyr
\item A character string naming a table/view in \code{conn}
}
Data is returned from this object.}

\item{by_feature}{Logical. If \code{TRUE}, the geometric operation is applied
separately to each geometry. If \code{FALSE}, the geometric
operation is applied to the data as a whole.}

\item{conn}{A connection object to a DuckDB database. If \code{NULL}, the function
runs on a temporary DuckDB database.}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
A logical vector
}
\description{
Functions to check whether geometries have Z (elevation) or M (measure) dimensions
}
\details{
These functions check for additional coordinate dimensions beyond X and Y:
\itemize{
\item \code{ddbs_has_z()} checks if a geometry has Z coordinates (elevation/altitude values).
Geometries with Z dimension are often referred to as 3D geometries and have
coordinates in the form (X, Y, Z).
\item \code{ddbs_has_m()} checks if a geometry has M coordinates (measure values). The M
dimension typically represents a measurement along the geometry, such as distance
or time, and results in coordinates of the form (X, Y, M) or (X, Y, Z, M).
}
}
\examples{
\dontrun{
## load packages
library(dplyr)
library(duckspatial)

## create a duckdb database in memory (with spatial extension)
conn <- ddbs_create_conn(dbdir = "memory")

## read data
countries_ddbs <- ddbs_open_dataset(
  system.file("spatial/countries.geojson", 
  package = "duckspatial")
) |> 
  filter(ISO3_CODE != "ATA")

## check if it has Z or M
ddbs_has_m(countries_ddbs)
ddbs_has_z(countries_ddbs)
}

}
