% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_filter.R
\name{ddbs_intersection}
\alias{ddbs_intersection}
\title{Calculates the intersection of two geometries}
\usage{
ddbs_intersection(conn, x, y, crs = NULL)
}
\arguments{
\item{conn}{a connection object to a DuckDB database}

\item{x}{a table with geometry column within the DuckDB database. Data is returned
from this object}

\item{y}{a table with geometry column within the DuckDB database}

\item{crs}{the coordinates reference system of the data}
}
\value{
an sf object
}
\description{
Calculates the intersection of two geometries, and return a \code{sf} object
}
\examples{
1 + 1 ## TODO

}
