% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbs_geom_validation.R
\name{ddbs_is_simple}
\alias{ddbs_is_simple}
\title{Check if geometries are simple}
\usage{
ddbs_is_simple(
  x,
  conn = NULL,
  crs = NULL,
  crs_column = "crs_duckspatial",
  quiet = FALSE
)
}
\arguments{
\item{x}{Input spatial data. Can be:
\itemize{
\item A \code{duckspatial_df} object (lazy spatial data frame via dbplyr)
\item An \code{sf} object
\item A \code{tbl_lazy} from dbplyr
\item A character string naming a table/view in \code{conn}
}
Data is returned from this object.}

\item{conn}{A connection object to a DuckDB database. If \code{NULL}, the function
runs on a temporary DuckDB database.}

\item{crs}{\link{Deprecated} The coordinates reference system of the data. Specify if the data
doesn't have a \code{crs_column}, and you know the CRS.}

\item{crs_column}{\link{Deprecated} a character string of length one specifying the column
storing the CRS (created automatically by \code{\link{ddbs_write_vector}}).
Set to \code{NULL} if absent.}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
A logical vector
}
\description{
Determines whether geometries are simple. That is, free of self-intersections and
returns a boolean indicator of simplicity.
}
\examples{
\dontrun{
## load package
library(duckspatial)

# create a duckdb database in memory (with spatial extension)
conn <- ddbs_create_conn(dbdir = "memory")

## read data
argentina_ddbs <- ddbs_open_dataset(
  system.file("spatial/argentina.geojson", 
  package = "duckspatial")
)

## store in duckdb
ddbs_write_vector(conn, argentina_ddbs, "argentina")

## check simplicity
ddbs_is_simple("argentina", conn)

## check simplicity without using a connection
ddbs_is_simple(argentina_ddbs)
}
}
