% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbs_join.R
\name{ddbs_join}
\alias{ddbs_join}
\title{Performs spatial joins of two geometries}
\usage{
ddbs_join(
  conn,
  x,
  y,
  join = "ST_Intersects",
  name = NULL,
  crs = NULL,
  crs_column = "crs_duckspatial",
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{conn}{A connection object to a DuckDB database}

\item{x}{a table with geometry column within the DuckDB database. Data is returned
from this object}

\item{y}{a table with geometry column within the DuckDB database}

\item{join}{A geometry predicate function. Defaults to \code{"ST_Intersects"}. See
the details for other options.}

\item{name}{a character string of length one specifying the name of the table,
or a character string of length two specifying the schema and table names. If it's
NULL (the default), it will return the result as an \code{sf} object}

\item{crs}{The coordinates reference system of the data. Specify if the data
doesn't have a \code{crs_column}, and you know the CRS}

\item{crs_column}{a character string of length one specifying the column
storing the CRS (created automatically by \code{\link{ddbs_write_vector}}).
Set to \code{NULL} if absent}

\item{overwrite}{Boolean. whether to overwrite the existing table if it exists. Defaults
to \code{FALSE}. This argument is ignore when \code{name} is \code{NULL}.}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
an sf object or TRUE (invisibly) for table creation
}
\description{
Performs spatial joins of two geometries, and returns a \code{sf} object
or creates a new table
}
\details{
Spatial Join Predicates:

A spatial predicate is really just a function that evaluates some spatial
relation between two geometries and returns true or false, e.g., “does a
contain b” or “is a within distance x of b”. Here is a quick overview of the
most commonly used ones, taking two geometries a and b:
\itemize{
\item \code{"ST_Intersects"}: Whether a intersects b
\item \code{"ST_Contains"}: Whether a contains b
\item \code{"ST_ContainsProperly"}: Whether a contains b without b touching a's boundary
\item \code{"ST_Within"}: Whether a is within b
\item \code{"ST_Overlaps"}: Whether a overlaps b
\item \code{"ST_Touches"}: Whether a touches b
\item \code{"ST_Equals"}: Whether a is equal to b
\item \code{"ST_Crosses"}: Whether a crosses b
\item \code{"ST_Covers"}: Whether a covers b
\item \code{"ST_CoveredBy"}: Whether a is covered by b
\item \code{"ST_DWithin"}: x)	Whether a is within distance x of b
}
}
\examples{
\dontrun{
## load packages
library(duckdb)
library(duckspatial)
library(sf)

## database setup
conn <- duckdb::dbConnect(duckdb::duckdb())
ddbs_install(conn)
ddbs_load(conn)

## read data
countries_sf <- sf::st_read(system.file("spatial/countries.geojson", package = "duckspatial"))
argentina_sf <- sf::st_read(system.file("spatial/argentina.geojson", package = "duckspatial"))
brazil_sf <- subset(countries_sf, NAME_ENGL == "Brazil")

## store in duckdb
ddbs_write_vector(conn, countries_sf, "countries", overwrite = TRUE)
ddbs_write_vector(conn, argentina_sf, "argentina", overwrite = TRUE)

## spatial join: intersects
temp_1 <- ddbs_join(
    conn, x = "countries",
    y = "argentina",
    join = "ST_Intersects"
    )

head(temp_1)

}
}
