% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbs_utils.R
\name{ddbs_list_tables}
\alias{ddbs_list_tables}
\title{Check tables and schemas inside a database}
\usage{
ddbs_list_tables(conn)
}
\arguments{
\item{conn}{A connection object to a DuckDB database}
}
\value{
\code{data.frame}
}
\description{
Check tables and schemas inside a database
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
## TODO
2+2
\dontshow{\}) # examplesIf}
}
