% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{ddbs_list_tables}
\alias{ddbs_list_tables}
\title{Check tables and schemas inside a database}
\usage{
ddbs_list_tables(conn)
}
\arguments{
\item{conn}{A \code{DBIConnection} object to a DuckDB database}
}
\value{
\code{data.frame}
}
\description{
Check tables and schemas inside a database
}
\examples{
\dontrun{
## load packages
library(duckspatial)

## create a duckdb database in memory (with spatial extension)
conn <- ddbs_create_conn(dbdir = "memory")

## read some data
countries_ddbs <- ddbs_open_dataset(
  system.file("spatial/countries.geojson", 
  package = "duckspatial")
)

argentina_ddbs <- ddbs_open_dataset(
  system.file("spatial/argentina.geojson", 
  package = "duckspatial")
)

## insert into the database
ddbs_write_table(conn, argentina_ddbs, "argentina")
ddbs_write_table(conn, countries_ddbs, "countries")

## list tables in the database
ddbs_list_tables(conn)
}
}
