% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_read.R
\name{ddbs_read_vector}
\alias{ddbs_read_vector}
\title{Load spatial vector data from DuckDB into R}
\usage{
ddbs_read_vector(conn, name, clauses = NULL, quiet = FALSE)
}
\arguments{
\item{conn}{A \code{DBIConnection} object to a DuckDB database}

\item{name}{A character string of length one specifying the name of the table,
or a character string of length two specifying the schema and table
names. If \code{NULL} (the default), the function returns the result as an
\code{sf} object}

\item{clauses}{character, additional SQL code to modify the query from the
table (e.g. "WHERE ...", "ORDER BY...")}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
an sf object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{ddbs_read_vector()} was renamed to \code{\link{ddbs_read_table}}.
}
\keyword{internal}
