% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_register.R
\name{ddbs_register_vector}
\alias{ddbs_register_vector}
\title{Register an SF Object as an Arrow Table in DuckDB}
\usage{
ddbs_register_vector(conn, data, name)
}
\arguments{
\item{conn}{A connection object to a DuckDB database}

\item{data}{A \code{sf} object to register in the DuckDB database.}

\item{name}{The name of the temporary view to create in the database.}
}
\value{
TRUE (invisibly) on successful registration.
}
\description{
This function registers a Simple Features (SF) object as a temporary Arrow-backed
view in a DuckDB database. This is a zero-copy operation and is significantly
faster than \code{ddbs_write_vector} for workflows that do not require data to be
permanently materialized in the database.
}
\examples{
\dontrun{
library(duckdb)
library(duckspatial)
library(sf)

conn <- ddbs_create_conn()

nc <- st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)

ddbs_register_vector(conn, nc, "nc_arrow_view")

dbGetQuery(conn, "SELECT COUNT(*) FROM nc_arrow_view;")

dbDisconnect(conn, shutdown = TRUE)
}
}
