% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_write.R
\name{ddbs_write_vector}
\alias{ddbs_write_vector}
\title{Write an SF Object to a DuckDB Database}
\usage{
ddbs_write_vector(
  conn,
  data,
  name,
  overwrite = FALSE,
  temp_view = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{conn}{A \code{DBIConnection} object to a DuckDB database}

\item{data}{A \code{sf} object to write to the DuckDB database, or the path to
a local file that can be read with \code{ST_READ}}

\item{name}{A character string of length one specifying the name of the table,
or a character string of length two specifying the schema and table
names. If \code{NULL} (the default), the function returns the result as an
\code{sf} object}

\item{overwrite}{Boolean. whether to overwrite the existing table if it exists. Defaults
to \code{FALSE}. This argument is ignored when \code{name} is \code{NULL}.}

\item{temp_view}{If \code{TRUE}, registers the \code{sf} object as a temporary Arrow-backed database 'view'
using \link{ddbs_register_table} instead of creating a persistent table. This is much faster but the view
will not persist. Defaults to \code{FALSE}.}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
TRUE (invisibly) for successful import
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{ddbs_write_vector()} was renamed to \code{\link{ddbs_write_table}}.
}
\keyword{internal}
